update setTopBlock to long as getTopBlock is also using long and blockStore.getLastBlock().getBlockNum() also. update pom profiles

+++ src/main/java/io/smilo/commons/peer/sport/INetworkState.java
@@ -13,7 +13,7 @@ public interface INetworkState {
 
     long getTopBlock();
 
-    void setTopBlock(int topBlock);
+    void setTopBlock(long topBlock);
 
     Set<Network> getNetworks();
